requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
streamlit==1.29.0
plotly==5.17.0
python-dateutil==2.8.2
aiohttp==3.9.1
asyncio-throttle==1.0.2
openpyxl==3.1.2
schedule==1.2.0 